Work in progress.
Working on another video.
Added a first draft of the script.
Added some images.

Working on a way to save the SVG that I create as an SVG file.
I'm using the link icon as an example of a working SVG.
That will be replaced by the graphics at the top.