First version with morphing text.
Lots of TODO items.
But this is a good sign that I'm on the right track.

Totally reorganize path-shape.ts.
Now each command knows it's own starting position.
So it can be manipulated more easily.
A command does not need a pointer back to the previous command any more.
PathBuilder keeps a history of commands to make it easier to create paths.
I.e. it fills in the starting values for you.