Added curves.html
This includes some interesting ways to create and manipulate paths.

I actually created the curves demo before the rest of this project.
Curves.html was in its own repository.
I never bothered to publish it to GitHub.
I forgot about it until just now.
It made more sense to add it to this repository than to publish the original repository.