Change the B to look better when rough.
It's the same shape.
But I drew it in a different order.
So I don't have to lift the pen so often.
So the rough code has more context when randomizing stuff.
I.e. the code will smooth out more of the connections between segments.

The range control now returns a value between 0 and 1.
It was going down to -0.0001 to fix a bug in the handwriting animation.
Now that fix has moved into the handwriting animation, where it belongs.

Bugs:

You can see some of the known TODOs better now.
Notice how there are a lot of kinks where things should be smooth.

The rough animation  is supposed to translate smoothly the original and the rough version.
But all the way to the left it's still a little rough!