Improvements related to Q commands.
Add a lot more to the path-debugger.*.
That is how I found some of the bugs that I fixed.

Fix some of the "rough" code in sky-writing.*.
Sometimes a Q command's angles are not exactly as requested.
Each QCommand now stores a copy of the original request that created it.
Displaying the requested angles in the path-debugger.* helped find some issues.
Actually using this data fixed other issues.