More upgrades to the path debugger.
When you click on a rough character in sky-writing.html it automatically copies the debug string.
Previously it wrote that to the console and you had to copy it yourself.
And I've made corresponding changes to path-debugger.html.
You can usually just hit ⌘C and return to display the result.

The "Processed" section of path-debugger.html now uses the selection to highlight important things.
The user still clicks on segments in the verbatim section to highlight them.