Lots of improvements to the rough code in sky-writing.html.
Limit the randomness based on how large each segment is.
Remove the old `scale` option.
This helps a lot, but it will not solve all the problems.

Add some attempts to fix bad curves.
When the first pass creates unexpected kinks and flat spots,
now we make a second pass to try to clean up.
And the code is organized to allow more passed to be added.
This helps a lot in specific cases.
I need to do more work here.

Add code to split a single Q command into two smaller Q commands.
This is required when the number of curves in the before and after cases aren't the same.
Currently this is used in exactly one place, but it will be very useful in a lot of places.

Fix the debugger in sky-writing.html.
The rough version was correct.
The other version was showing the original request.
Now it shows the morphable and hopefully similar version that was created to go with the rough version.
letter.shape is the original. letter.shape1 should look similar to the end user but will only use Q commands.

Bug fix in path debugger widget:
If you displayed - or | the previous revision would get confused.
It would try to set the width or height of the svg.viewbox to 0.
Then the SVG would display nothing.