Update a d g p q and b.
I redesigned these to look better when rough, but they look good normal, too.
In each of these cases I was lifting my pen between the loop and the straight part.
Now I stretch the end of the loop to meet the end of the straight part.
I've tried this before, but I never made it look right.
I finally got the `extra` part right so it looks great now.