Add testCircleApproximation() to sky-writing.ts
This will send test data to the console.
This will show the incoming and outgoing angles of each command in certain characters.
These are a little off at the edges of each circle or circle segment.
I plan to change the approximation and this will show if my fix helps or not.