Test an image tag inside an SVG.
This seems to work better in some circumstances that others.

This is how to view the SVG by itself:
https://raw.githubusercontent.com/TradeIdeasPhilip/random-svg-tests/master/public/sample-1.svg

This is how to see it embedded in a markdown file on the web:
https://github.com/TradeIdeasPhilip/random-svg-tests/blob/master/svg-for-programmers.md

I tried looking at the preview for that page in VS Code.
It never gave me the results I expected.