tangent-line-2 ready for release.
I added a resize observer to fix a bug.
Sometimes the colors were wrong on the legend.
It was rare in interactive testing, but it happened every time on video.

I removed the circle hiding animation.
First I changed it, but then I commented it all out.
It didn't add much, and it was too distracting.
It might have worked if things were moving much more slowly.

Now I hide the debug interface before recording the moving.

See the result here:  https://youtu.be/0pZ10xOPHL4