The dx demo video is now complete.
This now covers the entire voiceover.

I added spotlights to the beginning of the script.
Those match the existing voiceover.
This commit finishes the spotlights, including smooth transitions.

I removed a lot of obsolete code.
I tried a lot of different things as I was building this project.
Some of the old debug tools make no sense anymore.
And numbers are now stored as numbers.  😁

I used this commit to create https://youtu.be/THZZlEpo684