Final version used in video.
Make the dotted lines overlap a little.
When they are all lined up perfection, I used to see some gaps.
Now I see blurry areas in their place.

Adjust the blurry part of the script to match the voiceover.

The script is now 8 seconds longer.
The voiceover includes a long pause between this animation and the next one.
This video used to end at the at beginning of the pause.
Now it goes half way into the pause.

Add zooming.
This was always planned.

Add a nice stop after finishing the realtime animation.
This would cause the debugger to stop everything.
Or fill the log with error messages.
Now it just stops at the end of the script.
(Asking for a frame before the beginning or after the end of the script will throw an error.)