Several improvements to the GUI.
We automatically run the default function without waiting for the user to do anything.

We can now have more input sliders.
We start with 2, but the user can request more.

The syntax for accessing the sliders has changed slightly.
The default function has changed, accordingly.

Display error messages.
In a user friendly way in a red box.
Include a button sometimes to fix the problems.

Disable the GO button when appropriate.
Changing the code for the function does NOT immediately rerun anything.
It just reenables the GO button so it's ready when you are ready.

You can now set the segment count from the GUI.