Tweak the GUI.
I moved the inputs section closer to the output samples.
So you can move a slider and immediately see the result without scrolling.

I fixed the margin.
The margin for <html> looked good in development mode.
But it went to 0 in production.
I'm not sure of the underlying problem.
I adding !important to fix it.