Refactor code.
Minimal changes to the user interface.
The SVG used to help with clipping and masking is now explicitly visible.
This may help us with debugging.
Otherwise things look the same to the end user.

A lot of calls to createElementNS() have been replaced with simpler code in *.html and *.css.