Update the animations in the table.
The numbers were updating at the wrong times.

If a row has no circles, draw it with low opacity.

When the data in a row is changing,
making it briefly wink out and back in,
using the opacity.

Clean up the console log.
Remove anything that didn't have to do with the animations.
Clean up up the logs for the animations.