Add more terms to the HiddenPoles example.
This made a small problem become a big problem.
Near the radius of convergence the y values oscillate between very large and very small numbers.

Add easing to the way that we display Taylor series.
So we take more samples near the extremes of the range.
Because we know that the function goes crazy around the extremes of the range.
And I want to display that right, especially the transition between accurate and crazy.