Make the open endpoints hollow.
Previously they were black, to match the background color.
That was a quick approximation of what I really wanted.
I had to add a mask to cover the path attached to each circle.
Otherwise the path would be visible inside the circle.

And remove some ease() code from the previous commit.
On closer inspection it didn't help and in some cases it made things slightly worse.