Add circle at the ideal end points.
The open endpoints sometimes move around a lot, and go to way off into the weeds.
So draw an additional circle where this point should be, according the to the original function.
Draw the new circle with a dotted border, so it's clearly less important than the real circle.
And mask out any part of the ideal circle that would have been behind the real circle.