Add an option to hide or show the bad points.
That includes the radius of convergence, etc.
This is turned off in the two existing scripts.
This wil be turned on for the second half of the video.
You can use a checkbox in the #manual-controls to show or hide this.

This feature is completely working now.
It is ready for use in the next script.