parametricPath() now works with a constant parametric function.
Bug fix.
In the past if a function was constant, always returning the same value, that would lead to a failed assertion.
Now we recognize that case and deal with it appropriately.
A constant function is not that interesting on its own.
But it can be the first step in an animation.
You can now morph from the constant function to any other function without issue.