Even out the animations.
We were showing the circles one at a time.
If we had too many circles, all of the extras went into the same bin.
All bins but the last had exactly one circle.
Now we can combine the circles in more ways.
We try to make all of the steps as similar as possible.

There was a big visible jump at the end of some animations.
Now the animations move more smoothly.