Add "like share and subscribe."
You can see this at the bottom of hershey-fonts-viewer.html.
You can see this if you select "SVG Path" at complex-fourier-series.html

Note that I reversed every other line.
So the text goes back and forth, instead of each line going left to right.
To make it easier on the Fourier series.
When it has to add lines to connect things, the lines will be shorter and won't hide as much.