path-to-fourier has started in earnest.
Add error checking to complex-fourier-series
If the sampleCount is too low you'll see pink and dark red lines.
Use support.sampleCount to fix the problem.

Rationale:  I wanted to know how stable this path was.
If I make small tweaks to the path, will that cause instabilities?
See the new plan / TODO at the bottom of path-to-fourier.ts.