Recenter many examples.
So we start with a circle instead of a line.
Or at least we start with a smaller or more aesthetically pleasing line.

Bug fix.
In some cases the new PathShape.parametric1() function would go infinite recursion.
(That function is now called glitchFreeParametric().)
This was causing the web page to lock up from time to time.
It was especially bad when first transitioning from a point to a circle.
Sometimes, at the very beginning of that animation there would be a path with exactly one segment and very little length.
This often causes trouble.