Added startAt slider.
And the segmentCountInput updates better now.
It used to only update when you let go.
I don't remember if that was on purpose, or a mistake.
Now it updates instantly, just like the other controls.

ε becomes 100 times smaller in PathShape.parametric().
The previous value seemed big when looking at real examples.

Add new examples:
Cusps and Squares.
Both push my ability to deal with sudden turns.