Internal release.
I'm in the middle of changing a lot of things.
It will be useful to have https://tradeideasphilip.github.io/random-svg-tests/fourier-smackdown.html?index=11 available.
Previously you had to run this in the debugger to view it.

I'm working with https://github.com/TradeIdeasPhilip/html-to-video
That project and this web page are evolving together.