Internal release.
Preparing for a new video.
I want to synchronize the three examples.
I've made some changes.
And this is a good place to commit.

Return to the red, white and blue color scheme.
Solid black background.
15 spinners are visible again.

Reorganize the code.
I'm no longer calling createScript() / groupTerms() from fourier-shared.ts.
That code was ugly.
And this will allow me to tweak more things in the future.
Currently it works mostly like before.
But it doesn't group any circles (it could, it's just not configured to).
And it's stopping at 15 circles max, fewer than before.