Bug fix:
All of the numbers were wrong.
In the previous commit I removed the "background" copy of the text.
(That never did what is should, and recently it was causing issues.)
And I broke the main text in the process.
Everything said “-3” because that was the default built into the HTML.

I have restored the correct text.
And the offending "background" text remains banished.